+++ packages/slipcast/config/browsersync.js
@@ -3,11 +3,11 @@
 const config = require('./slipcast');
 const { join } = require('path');
 
-module.exports = {
+module.exports = Object.assign({
   files: [
     join(config.output, '**', '*.{css,jpg,js,png,svg}'),
     join(config.folders.static, '**', '*'),
   ],
   open: false,
   server: [config.output, config.folders.static],
-};
+}, config.watch.browsersync);